Fix filtered search token specs
The feature to add navigation by arrow keys in `GlFilteredSearch`,
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/2752, added
a new required prop `cursorPosition` in `GlFilteredSearchToken`.

Some implementations of custom token components here needed to be
changed to pass `$attrs` down to the underlying `GlFilteredSearchToken`
to satisfy this new prop requirement, and tests needed to be updated as
well to provide that prop at all. In reality, that prop is provided
automatically by `GlFilteredSearch`; tests needed to be changed because
they artificially instantiate the token components *without* a real
ancestor `GlFilteredSearch` instance.