Limit to current PG user
When exporting query statistics from pg_stat_statements,
we're only interested in those for the current PG user.
This is the user that also runs database migrations, for example.

If we exported all and didn't have enough privileges, we'd
start to see query strings with "<insufficient privileges>" instead.

We use a slightly odd construct here to work around a permissions
problem on .com:

```sql
pg_get_userbyid(userid) = current_user
```

This is because we don't have privileges on `pg_authid`.

Changelog: other