Update tests to reflect changed results
The change is that for resized images, we now return URLs that include
version SHAs which are relevant to that actual design (rather than a SHA
at which it was not modified). This does not affect the image itself.

So now any expectation on the image URLs needs to consider the actual
version.