Add requestAttention/removeAttentionRequest mutations
This adds mutations (`mergeRequestRequestAttention` and
`mergeRequestRemoveAttentionRequest`) to request attention of a user
and to remove attention request for a certain user. This will be
used by the frontend in place of `mergeRequestToggleAttentionRequested`.

This is the backend part of fix for
https://gitlab.com/gitlab-org/gitlab/-/issues/357349.

This is still behind the `mr_attention_requests` feature flag.