Separate Sidekiq metrics and health-checks server
It has been a long-standing issue that for Sidekiq,
serving metrics and health-checks was coupled into
a single server.

We had already separated settings keys for these two
concerns previously, in anticipation of splitting
these endpoints apart, which is a breaking change.

We can now do this with 15.0.

Changelog: changed