Support for TLS config on Workhorse
As part of a requirement to use end-to-end encryption, Workhorse needs
to be able to run HTTPS server.

It allows to specify [tls] config in config.toml.
The config must contain paths to certificate and private key.
After that, the server will be accessible via HTTPS.

Relates to #353010

Changelog: added