git: Parse and display gitaly limit error
Gitaly will soon return errors of the type LimitError. These can occur
whenever Gitaly determines that it is overloaded and can't keep up with
the current request load. It is a way for Gitaly to impose backpressure
on clients.

When this error gets returned to workhorse, we want to recognize it and
write something meaningful that Git recognizes and can pass through to
the end user.

This commit changes how we handle errors for upload-pack and
receive-pack.

Changelog: changed