Stops semantic diff if file is truncated
Ipynb semantic diffs need to look at the entire file to generate the
diff, but when the blob is too large it comes truncated and it will
break. We could load the entire data, but then it would hit the
timeouts, so we would just do more computation for nothing.