Clarify cluster deletion behavior
All of the cluster deletion APIs only
delete the linked cluster object in GitLab
and do not invoke the cleanup functions
that's available as an option through
the UIs.

This change adds sentences to all 3
APIs that delete clusters to clarify
their behavior w.r.t. underlying resources
within the connected Kubernetes clusters.