Add pipeline tabs redirection support
When the feature flag pipeline_tabs_vue is on,
we now use GlTabs instead of haml ones in
the pipeline page. We want to support two usecase:
the legacy one had each tab on its own route, so
now we redirect the user to the right tab if that's
the case. The other use case is by using query params.