Update full-text search index when moving issues
When moving issues, we do several updates to the new issue within an
outer transaction. This caused `saved_changes` to return only the
changes of the latest update even if more changes were committed.

This meant `#update_search_data!` wasn't run even though we saved the
title and description of the issue.