Show deployment approval on deployments page
This is the beginning of showing deployment approvals historically, as
it will show the approval status of all pending, finished, and failed
deployments. It does not yet show comments regarding the deployments,
only who has approved or rejected the deployment, and how long ago it
was.

Changelog: added
EE: true