Add status field to Rack Attack log messages
When Rack Attack (or another middleware using the same log format)
rejected a request, we were not including a status code. This meant that
breaking down request logs by status would be misleading, and
under-report 403 and 429 responses, as those are the two statuses
returned by Rack Attack.

This was confusing and has come up in a couple of production incidents.
We now include the relevant status code in these messages when Rack
Attack is stopping the rest of the request from processing. When it
isn't - in dry-run mode - we don't add a status field, as then we'd have
the opposite problem and double count with the regular request logs.

Also remove some unused constants.