Start storing changed_at in seat statistics worker
We want to know the date that the max_seats_used was last increased so
that we can efficiently alert the user about upcoming charges due to
changes in seat usage. Update the statement in the seat statistics
refresh worker as this statement bypasses the before_update callback on
the model.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/348481
Changelog: added