Frontend for Environment Search
We use the GlSearchBoxByType for instant searching, piping the query
value through the local graphql resolvers to send off to the backend.

For some reason, if we don't call resetPolling when changing the
variables, apollo doesn't listen to the polling interval, and instead
spams network calls over and over and over, slowing the page to a crawl.

Changelog: added