Update runner status summary upon filtering
In order for users to see more up to date information about their
runner, this change has the runner stats match the current filters
that have been selected by the user.

It moves the runner stats summary block below the search bar so the
visual hierarchy of the filters is clear.

Changelog: changed