gitaly_client: Allow filtering quarantined commits by existence
The `#list_new_commits` function can optionally use the quarantine
directory of a repository to enumerate commits. Instead of doing a
graph-walking and walking all existing revisions to figure out what
objects are new, we then instead enumerate all object part of the
staging environment directly. This is a lot faster in repositories which
have a lot of revisions.

This way of doing things has a downside though: when a client pushes
objects, then this push may also contain objects which already exist in
the target repository. Consequentially, the quarantine directory will
also contain these objects, and because we directly enumerate and return
all such staged objects the returned set of supposedly new commits may
also contain commits which in fact already exist. This has the potential
to break some access checks, and most prominently it breaks the author
checks.

To fix this issue, Gitaly has implemented a new `CheckObjectsExist()`
RPC. Given a set of revisions, Gitaly will determine for each of them
whether they exist or not in a specific repository. Equipped with this
RPC, we can now filter down the set of quarantined commits to the set of
actually-new commits by asking the repository whether it knows the
quarantined objects already. Note though: we must make sure that the
repository we send as part of this query does not contain any reference
to the quarantine directory, or otherwise we'd wrongfully claim that all
quarantine objects exist already.

Prepare for a fix by implementing this filtering in `#list_new_commits`
if the caller passes a new `filter_quarantined_commits` parameter.