single_change_access: Use quarantine directory to find new commits
While the `ChangesAccess` class already uses the quarantine directory to
find new commits, the `SingleChangeAccess` class doesn't. And ultimately
it doesn't matter: the only callsite which instantiates this class is
`ChangesAccess` itself, and it already passes in the set of new commits.
So in production, there is no case where the `SingleChangeAccess` would
actually try to find new commits by itself.

Migrate `#commits` to use the quarantine directory regardless of that
fact. While it doesn't make any difference for us, it at least ensures
that all callsites of `#new_commits` are called with quarantine
directories enabled. When we get rid of the feature flag we can thus get
rid of the parameters which control this.