changes_access: Do not recompute whether quarantine can be used
While we already cache the result of `#commits`, we still recompute the
parameters required to compute the result every time. While this is
likely not a performance problem in general given that we only loop over
all changes, it still is needlessly wasting compute cycles.

Fix this by using the `strong_memoize` helper.