changes_access: Speed up commit-based access checks for new refs again
With d86514f1068 (checks: Fix revalidation of preexisting commits,
2021-12-14), we had to disable use of the quarantine directory to
find new commits for our access checks in case any pushed reference has
a blank old revision, which is the case for all branch creations. The
reason for this is that the quarantine directory may contain commits
which indeed exist in the target repository already, and if we didn't
have an old revision we couldn't find the set of new commits by doing an
in-memory graph walk because we had no cut-off point.

As was known back then, this change is a huge performance regression: if
we don't use the quarantine directory, we instead use `git rev-list
--not --all $newrev`, which is forced to walk all references which exist
in the repository to mark preexisting commits as uninteresting. This of
course scales with both the number of commits and with the number of
references. In our own `gitlab-org/gitlab`, this takes around 23 seconds
to compute, which is plain unacceptable for our access checks. On the
other side, using the quarantine directory for this only takes about 5
milliseconds.

With the changes in the preceding commits we now have the infrastructure
to again switch back to using the quarantine directory via Gitaly's new
`CheckObjectsExist()` RPC. We simply filter down the list of quarantined
commits by whether they exist in the target repository or not. If they
do, there is no need to perform any access checks for them.

Start calling `#new_commits` with `filter_quarantined_commits: true` in
case a feature flag is set. As explained, this should now really only
return non-existing commits and result in a significant performance
boost in large-ish repositories.

Changelog: performance