Load group Vulnerability Report when only subgroups have projects
This changes the group Vulnerability Report's `has_projects` property
implementation so that it accounts for sub-groups. This fixes an issue
where, if a group did not have any projects but had sub-groups that did,
the Vulnerability Report would be stuck in the empty state because the
`has_projects` property evaluated to `false` as it only looked for
top-level projects.

Changelog: fixed
EE: true