Forbid recursion during Feature.enabled?
Any call that involves recursion evaluates to the default value.

This introduces with_feature to ensure that we pop the recursion stack
correctly in all the places where we call `Feature.unsafe_get`, which is
as the name indicates, a highly unsafe method. If this stack is not
cleared correctly, subsequent lookups will short-circuit.

Changelog: fixed