Re-add usage_data_diff_searches flag
It turns out that there are several feature flags that are dynamically
checked for during frontend stats collection. If the definition is
removed, we start sponting errors over into Sentry. This situation
therefore creates a class of "undead" feature flags that must live in
perpetuity in our codebase. We have an issue
(https://gitlab.com/gitlab-org/gitlab/-/issues/350713) for remedying
this situation, tentatively scheduled for 15.2.. until then, we must
keep flags like this one around as definitions, even if we're not
employing them anywhere in the code base.

Changelog: fixed