Allow Slack app to be reinstalled
To roll out future updates we have to provide a button on the
integration form to reinstall the app, which works the same as
installing the app for the first time.

This replaces the current "Add to Slack" PNG button with an SVG version,
and also displays the button with the label "Reinstall Slack app" when
the app is already installed.

In the backend we update the existing `SlackIntegration` record instead
of failing when it already exists.

We also start storing the bot user ID and access token, which we don't
use yet but will need for other features.

EE: true
Changelog: changed