Create service to calculate the seat count usage
We want to be able to display an alert to the user when they are nearing
their seat count for their subscription.

Create a service that will determine if a use should be shown this alert
based on a differing set of thresholds for their company size.

Issue: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/79563
EE: true