Marginalia - Supply lines in test env
Reverts an older test change now that we include line in
the test environment as well as in dev.

Updates the merge_requests_spec to read expected queries from the query
recorder log. We should be examining the `log` as the robust source for
this information. Other changes are to help with understanding and
tracing the test failure.