Delete repositories without renaming them first
Prior to this change, the service responsible of deleting a repository
would first rename the repo to +deleted.git and then schedule a worker
to delete that 5min later. This is annoying for Gitaly, and Gitaly
Cluster in particular, because renames are complicated when there are
multiple replicas.

To overcome inconsistency issues when handling a RemoveRepository RPC in
Gitaly Cluster, changes were made in gitaly@f328e3f70 (repository:
Implement atomic locking semantics for RemoveRepository, 2021-11-30).
This makes the rename step in the Repositories::DestroyService
unnecessary.

This change removes the rename step and relies on Gitaly to handle the
RemoveRepository in an atomic way.

Changelog: performance
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/37086