Allow watching not-yet-created key in Workhorse Redis pub/sub
Watching fails directly if the key does not exist. However,
in some cases we may want to keep waiting until the key
gets created (or time out as before).

One example is Geo waiting for the Geo specific key to be created,
as it would not exist if Geo was not enabled before (and/or not EE).

This adds a `watchMissingKey` argument to `redis.WatchKey` that
does **not** return the redis.ErrNil error if the key does not exist.