Remove DB checks from Geo secondary check in user routes
We need to know whether the current site is a secondary or not
to set the default Devise routes (that the app uses to redirect,
create paths to etc).

Previously, this used the general check of GeoNode.exists? and
verified that the current_node in the database has the secondary flag.

However, this ran queries in initializers, and we disallow that
as it won't work for upgrading to Rails 7 either.

This now uses the `geo_database_configured?` check, on non-Geo nodes
or Geo primaries, this should not exist, as there is no tracking DB
and Omnibus / CNG don't add this config, while they add it for
Geo secondaries.