Add an empty newline test to Ansi2Json
The current behavior of Ansi2Json when it encounters an
empty line is to ignore it in the JSON generated.

However, this behavior hasn't been explicitly tested
in the unit tests so this change adds a spec for it.

Regardless of this behavior being intentional or a bug
a test should be covering it given its observed during
use and a change to the behavior must not go unnoticed.

Changelog: other