Format dates as UTC
Group - Settings - Usage Quotas:

- Pipelines tab
- Shared runner usage tab

In the month data point charts displayed on
both of the above pages the backend sends us
UTC dates representing start of month such as
2022-01-01.

When transforming these into an `mmm yyyy`
format within the frontend we do not explicitly
mention to the formatting library that they need
to be treated as UTC dates so the browser's
timezone environment instead influences the result.

For timezones in the 'west', or anything that's
GMT minus n the formatted month becomes incorrect
and shows the prior month because of the negative
offset applying on the first second of the date.

This change forces the formatting to treat the dates
passed as UTC and adds tests that fail without the
fixes added.

Changelog: fixed
EE: true