Persist group name and use path for destination name
This change adjusts the transformation of group and sub-group data
when importing a group from one GitLab instance to another using the
bulk importer. It persists the source group/sub-group name and derives
the group/sub-group path from the destination name. This ensures the
group name is titleized and the path parameterizes.

If a diffferent group name is required, the user can change the name
of the newly imported group manually.

Changelog: changed