Fix .tree-table class not applying
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/84683,
we introduced some styles to clean up `GlTable`s in
the GitLab code. This caused a regression with the
`.tree-table` due to CSS specificity of the class
(lower than `table.table:not(.gl-table)` which includes `not`).

This MR adds the `table` tag to the CSS styles to
make sure they apply properly to the tree table.