Close issues async after merging MR
When closing issues errors out, it can stop the execution of
`PostMergeService` which can result to other operations not being
executed.

To ensure that we still run the rest of logic in `PostMergeService`
when that happens, we close issues asynchronously.

This is behind the `async_mr_close_issue` feature flag as a new
worker is being introduced.