Remove support for Cluster Image Scanning Analyzer
We are removing support for the Cluster Image Scanning analyzer with
GitLab 15.0 (https://gitlab.com/gitlab-org/gitlab/-/issues/356465). This
change removes the CI templates which are used to run the analyzer, and
the security orchestration policies code which is used to create
pipelines.

Note that only a small subset of Cluster Image Scanning functionality is
being removed. The Cluster Image Scanning feature itself remains a part
of GitLab.

- Cluster Image Scanning can still be used via the GitLab Agent
- The cluster_image_scanning job artifacts and vulnerability reports
  types will remain. Custom integrations can still output
  cluster_image_scanning reports.

Changelog: removed
EE: true