Remove ImportStateNotFound from ProjectImportScheduleWorker
We cannot recover from this error, so we don't need to retry it.

Retries were previously disabled for this worker, counting on it to be
rescheduled again. But we want to be able to retry this job so we can
allow it to use replicas, and can recover from an error if a replica
isn't caught up.