Force ImportScheduleWorker to use read-replica
This makes sure that the ProjectImportScheduleWorker uses replicas for
it's reads in the beginning of the job.

We need to allow retries for the `delayed` strategy, because that will
raise an error if it cannot find an up-to-date replica in 0.8s the
first time.

If the job can't find an up-to-date replica in the retry, it will use
the primary instead.