Change style of trigger job in pipeline graph
Trigger jobs (or bridges) are a special kind of
jobs in pipelines whose goal is to start a separate
pipeline either for a pipeline in a different repo
or a different CI configuration file inside the same
project. To help users understand and find these easier,
we change the style of the trigger jobs inside the graph
and add a badge.

Changelog: changed