Compare agent version against KAS version
On GitLab.com, the GitLab version is almost always ahead of the KAS
version, but it is actually the KAS version that we care about.

This is not a complete fix, because on GitLab.com, the KAS version is
not determined by GITLAB_KAS_VERSION because we don't have automatic
deployments for KAS yet. But it is a more precise estimate than the
GitLab version, and will be mostly correct once the KAS version is
determined from that file (due to incremental rollouts, canary etc,
there is never a guaranteed "one true version").

See https://gitlab.com/gitlab-org/gitlab/-/issues/361786

Changelog: fixed