Fix QA Geo spec address setting
Instead of defining the gitlab_address manually, use
define_gitlab_address_attribute to set the initialized flag
as well.

Without this, define_gitlab_address_attribute! would run again
and reset the address, because of that the URL will end up
as the default GDK one instead of the expected primary URL.