Downgrade auto-deploy-image to v2.25.0
In v2.26.0 we added support for the ingressClassName field and
proactively used it when the networking.k8s.io/v1/Ingress API was
available. The problem with this is that old ingress controllers don't
provide an IngressClass, causing deployments to fail on old controllers.
This likely affects a large number of users. See
https://gitlab.com/gitlab-org/cluster-integration/auto-deploy-image/-/issues/206

Changelog: fixed