Exclude date-range controls from loading state
This prevents date-range controls from being hidden when the
vulnerabilities over time card is loading new data. This ensures that
date-range buttons don't loose the focus state when clicking on one of
the options for the first time.
It generally makes sense not to hide this portion of the card while data
is being loaded because it only shows static data that doesn't depend on
the API response.

Changelog: changed
EE: true