feature_flags: Default-enable use of filtered quarantine commits
With 4aba970e69f (gitaly_client: Allow filtering quarantined commits by
existence, 2022-05-03), we have introduced code that enables us to use
the repo's quarantine directory in order to enumerate commits in access
checks. This can lead to a significant speedup when enumerating commits
in repositories which have a lot of references because the previous way
of enumerating commits would require us to walk each of the preexisting
refs so that we can find the commits that really are new.

We have rolled the accompanying feature flag without any observed issues
so far, so let's default-enable the feature flag.

Changelog: changed