Fix figuring out project path from the configuration
The Gitaly documentation currently refers to identfiying the project's
location from the config on the disk. The project's location is already
clear as one needs to have access to the repository in order to read the
config. The intention is to say one can discover which project this
repository belongs to by reading the project path from the repo's config,
for example 'gitlab-org/gitaly'. Fix this in the documentation.