Do not vary upload tempfile prefix
When accepting uploads for local storage, Workhorse would use a
temporary filename based on the filename of the uploaded file. This
has no value because the file is removed by the end of the request,
and it added complexity to Workhorse. This commit changes Workhorse to
always use the "gitlab-workhorse-upload" prefix.

Changelog: other