Swap FK web_hook_logs to web_hooks for LFK
Replaces the FK on `web_hook_logs.web_hook_id` with an index,
thus allowing web-hooks with large numbers of log entries to be
deleted without timeouts.

We are not using a
[loose foreign key](https://docs.gitlab.com/ee/development/database/loose_foreign_keys.html)
here, because the clean up is done automatically by the partitioning
mechanism, which does not retain log entries indefinitely.

Changelog: changed