Create ClusterEnabledGrant model and populates table
This new table is going to be used to temporary permit a set of users to
use the certificate-based-clusters feature, which allowing us to remove
for the majority of other users. On our first attemp we tried to do this
with FF only on the namespace leve. But the spectrum of users that need
this feature enabled for longer is too big. So big that toggling the FF
for all these actors is something that our FF framework wasn't ready to
handle. So checking the enabled users directly from a table will be much
simpler, faster and easy on memory consumption.

Changelog: other