Rename Projects::ServicesController and change routes
Merge all actions into `Projects::Settings::IntegrationsContoller`,
and rename the previous `#show` action to `#index` for consistency
with the admin and group controllers.

The URL paths for the member actions change from `/-/integrations/...`
to `/-/settings/integrations/...`.

The `#index` action stays in `/-/settings/integrations`.

Also rename `Projects::ServiceHookLogsController` to
`Projects::Settings::IntegrationHookLogsController`, since it's routes
are now in this namespace as well.

This is part of our rename from "services" to "integrations" in
https://gitlab.com/groups/gitlab-org/-/epics/2504

Changelog: changed