Allow version conflicts in commits migration
When the update_by_query task is launched, Elastic takes a snapshot
of the index. There are instances where documents are changed before
the update_by_query updates the document. This results in a version
conflict.

By setting conflicts=proceed we are telling Elasticsearch to keep a
count of conflicts but proceed instead of aborting and halting
the update.

Changelog: changed
EE: true