Concurrency for generic package publish API
The generic package publish API creates the package at first file
upload.

When multiple file uploads are initiated by a client in parallel
they will each race to be the one also creating the as-of-yet
non-existent package record in the database.

Only one such racing request will win and all others will fail
with a transaction error because they'll be viewed as attempting
to insert duplicate (non-unique) records.

The endpoint must therefore only be called serially to guarantee
success of multiple file uploads under a single name and version.

This change adds a documentation note under the pre-requisites
section with an attempt at explaining why.