Prevent an exception from being raised during invalid user registration
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/52543 an audit
event was added as an "after request hook", and expected a persisted
user. When a user is invalid there is no persisted record and so it
cannot be audited in this way.

The bug fix was to skip the auditing when the user is not persisted.

Resolves https://gitlab.com/gitlab-org/gitlab/-/issues/363002