Update Praefect's TLS documentation
Praefect's TLS documentation indicates using a Common Name is an
alternative to Subject Alternative Name. This doesn't work anymore
since Go 1.15. Update the documentation to call for always using the
SAN.

Omnibus configures a default value for the unencrypted listening
address. As such, removing the configuration option of the listener
is not enough to disable the listener. Call out how to disable the
listener in the config.