Add exec_insert_all write method to loadbalancer
This makes `exec_insert_all` on a connection look like a write to the
loadbalancer. Causing subsequent queries to stick to the primary.

Both [upserts][0] and [inserts][1] use `ActiveRecord::InsertAll` to
build the query, which is then executed on the connection using
`#exec_insert_all` [2].

Because the connection proxy did not know about this method, it fell
back to the primary, but without causing the session to stick
there. This meant that any insert done there would not be found when
querying for it immediately after.

[0]: https://github.com/rails/rails/blob/66073335585f04f2ed0f5ef930eb3c8955d50a6a/activerecord/lib/active_record/persistence.rb#L243
[1]: https://github.com/rails/rails/blob/66073335585f04f2ed0f5ef930eb3c8955d50a6a/activerecord/lib/active_record/persistence.rb#L124
[2]: https://github.com/rails/rails/blob/66073335585f04f2ed0f5ef930eb3c8955d50a6a/activerecord/lib/active_record/insert_all.rb#L35