Add filters by milestone, author, and assignee to task by type
The task by type internal API endpoints, used to produce the "Type of
work" chart, can now filter by author, by assignees, and by milestone.
This makes it consistent with the other charts on the same page.

Using the same convention as the other charts, the filters are specified
with the following URL parameters:

  - `milestone_title` - a string with the milestone title to filter for
  - `author_username` - the username of the author to filter for
  - `assignee_username` - it accepts a single username or multiple
    usernames (like
`assignee_username[]=username1&assignee_username[]=username2`) of
assignees to filter for.

Changelog: changed
EE: true