Ignore no-contructor-return violations
Ideally, these violations would be rewritten. For instance, instead of
the constructor returning a singleton, there could simply be an exported
`getInstance` function.

It's less clear what's going on in the `EditorInstance` class, but
I suspect there's a refactoring that makes it more natural. Returning
a proxy from the constructor seems surprising, since the "instance" you
get back won't actually be an instance of `EditorInstance`.