Add job data for pipeline_tabs_vue
In the effort to migrate the pipeline tabs to
Vue, we now pass the right data so that the jobs
tab can render. We also add the appropriate specs
and lazy load the tab to ensure that we only fetch
the data when the user click on the tab and not
on initial loading.