Ignore import/no-cycle violations
These cycles include dynamic `import()` calls within functions, which
means that this *should* be safe at runtime. But, import cycles are
still potentially confusing, especially when bundle splitting is
involved, so we mark these with ignore directives to flag that this is
undesirable.

See https://github.com/import-js/eslint-plugin-import/issues/2265 for
a discussion about this.