access_checks: Always enable use of quarantine directory
Remove the feature flag that guards the use of the quarantine directory
when enumerating new commits for access checks. With this feature flag
gone, enumerating new commits is going to be a lot faster in repos with
many references given that we don't have to walk the repository's object
graph anymore, but instead only have to take a peek into the quarantine
directory that contains all new objects.

With the feature flag gone, there are no callers of `#new_commits`
anymore which disable use of the quarantine directory. So let's also
drop that parameter while at it.

Changelog: performance