Make `with_restored_connection_stack` to give correct connection
The `with_restored_connection_stack` completely rewrites connection
stack. At least in case of `finalize_batched_background_migration`
we require to have a connection to pass lower layers. Making the
`with_*` to return connection allows us to pass re-mapped connection
that was previously set to `ActiveRecord::Base`.

This also tests the `#finalize_` that a correct connection stack is
used.