Fix handling for `GitlabSchemasValidateConnection` for early start
It appears that each connection needs to be adopted. However,
in it's initial phase (before being adopted) the Rails
might execute SQL queries to configure the connection settings
(collation, timeouts, etc.). This will happen before
the connection is actually assigned to the connection pool,
as if such connection will be failed to be initialized
it will be orphaned.