Stop overriding jh/Gemfile.lock with Gemfile
The override was precluding us from using bundle frozen, because on JiHu
related CI jobs, we have a different Gemfile and Gemfile.lock.

This was needed previously when were generating JiHu's Gemfile
dynamically based on the canonical Gemfile.

But now that JiHu has a static Gemfile, we don't need this anymore.
For more context see the discussion at

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/87053#note_940600438