Allow dynamic import expressions in specs
Dynamic imports with non-literal expressions for the path can cause
problems in the build process, so we disallow it in regular code.
In specs, there is no build process, so the lint rule need not apply.

See https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2322 for
why it's disallowed in regular code.