Patch ActiveRecord::Generators::Migration#configured_migrated_path
We can set the migrations_paths in the database configurations
to tell Rails where the migrations for the additional database
will live when using the --database flag. We use this setting
also to include the Geo post-deployment migrations path. The
problem is that the Rails migration generator joins the elements
within the array into one path while creating the migration file
which breaks the regular and post-migration generators for the
Geo database.

This modifies ActiveRecord::Generators::Migration#configured_migrated_path
to exclude any post-migration path, and always return only one
value while creating a migration file.

Changelog: fixed
EE: true